+++ fake_t_test.go
@@ -28,3 +28,19 @@ func (f *fakeT) FailNow() {
 	f.failed = true
 	runtime.Goexit()
 }
+
+func (f *fakeT) Fatal(args ...any) {
+	f.errors = append(f.errors, fmt.Sprint(args...))
+	f.failed = true
+	runtime.Goexit()
+}
+
+func (f *fakeT) Fatalf(format string, args ...any) {
+	f.errors = append(f.errors, fmt.Sprintf(format, args...))
+	f.failed = true
+	runtime.Goexit()
+}
+
+func (f *fakeT) Log(args ...any)                 {}
+func (f *fakeT) Logf(format string, args ...any) {}
+func (f *fakeT) Name() string                    { return "fakeT" }